Safari-609.1.11.1/LayoutTests/fast/repaint/obscured-background-no-repaint-expected.txt

Test that obscured animated gif does not trigger repaints. This test requires DRT.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS repaintRects is ""
PASS successfullyParsed is true

TEST COMPLETE

 
